Minor Fixes According to Review
